# haircutapiDocumentation.yaml
openapi: 3.1.0
info:
  title: haircutapi
  version: '1.0'
  summary: Haircut store api
  description: ''
  contact:
    name: Thinh Nguyen
    url: t9ngdu00@students.oamk.fi
    email: t9ngdu00@students.oamk.fi
servers:
  - url: 'http://localhost:4000'
paths:
  /items:
    get:
      summary: Get stores information
      tags: []
      responses:
        '200':
          description: Get all the haircut store
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemData:
                    type: array
                    items:
                      $ref: '#/components/schemas/ItemData'
              examples:
                example-1:
                  value:
                    itemData:
                      - items_id: '1'
                        name: Bazooni Barbershop
                        address: 'Asemakatu 15, 90100 Oulu'
                        rate: []
                        price: 5
                        opentime: '10:00'
                        contact_number: +84 384 568 983
                        description: string
                        discount_rate: 0
                        instagram: string
                        created_at: '2019-08-24T14:15:22Z'
                        update_at: '2019-08-24T14:15:22Z'
                        location:
                          longitude: 105.83552
                          latitude: 21.01066
        '404':
          description: Haircut store Not Found
      operationId: Get all stores
      description: Get all the baber store information.
    post:
      summary: Post a item
      operationId: Create a new store
      responses:
        '200':
          description: Store created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
              examples:
                example-1:
                  value:
                    items_id: '1'
                    name: Bazooni Barbershop
                    address: 'Asemakatu 15, 90100 Oulu'
                    rate: []
                    price: 5
                    opentime: '10:00'
                    contact_number: +84 384 568 983
                    description: string
                    discount_rate: 0
                    instagram: string
                    created_at: '2019-08-24T14:15:22Z'
                    update_at: '2019-08-24T14:15:22Z'
                    location:
                      longitude: 105.83552
                      latitude: 21.01066
        '404':
          description: Can not create store
      description: |
        Create item with store information
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemData'
        description: ''
  '/items/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Get a store
      tags: []
      responses:
        '200':
          description: Get the haircut store by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
              examples:
                example-1:
                  value:
                    items_id: '1'
                    name: Bazooni Barbershop
                    address: 'Asemakatu 15, 90100 Oulu'
                    rate: []
                    price: 5
                    opentime: '10:00'
                    contact_number: +84 384 568 983
                    description: string
                    discount_rate: 0
                    instagram: string
                    created_at: '2019-08-24T14:15:22Z'
                    update_at: '2019-08-24T14:15:22Z'
                    location:
                      longitude: 105.83552
                      latitude: 21.01066
        '404':
          description: Information with given id was mot found.
      operationId: Get store by id
      description: Get a single hari cut shop information
    put:
      summary: Edit a store
      operationId: Update the store
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
              examples:
                example-1:
                  value:
                    items_id: '1'
                    name: Bazooni Barbershop
                    address: 'Asemakatu 15, 90100 Oulu'
                    rate: []
                    price: 5
                    opentime: '10:00'
                    contact_number: +84 384 568 983
                    description: string
                    discount_rate: 0
                    instagram: string
                    created_at: '2019-08-24T14:15:22Z'
                    update_at: '2019-08-24T14:15:22Z'
                    location:
                      longitude: 105.83552
                      latitude: 21.01066
        '400':
          description: 'Error '
      description: |
        Edit item with store information
    delete:
      summary: Delete a store
      operationId: delete-items-id
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemData'
              examples:
                example-1:
                  value:
                    items_id: '1'
                    name: Bazooni Barbershop
                    address: 'Asemakatu 15, 90100 Oulu'
                    rate: []
                    price: 5
                    opentime: '10:00'
                    contact_number: +84 384 568 983
                    description: string
                    discount_rate: 0
                    instagram: string
                    created_at: '2019-08-24T14:15:22Z'
                    update_at: '2019-08-24T14:15:22Z'
                    location:
                      longitude: 105.83552
                      latitude: 21.01066
      description: |
        Delte a store
...
components:
  schemas:
    ItemData:
      title: Items
      type: object
      description: ''
      properties:
        items_id:
          type: integer
          example: '1'
        name:
          type: string
          example: Zuy Minh
        address:
          type: string
          example: 'Ngõ 46c Phạm Ngọc Thạch, 106 C3 tập thể, Trung Tự, Đống Đa, Hà Nội, Việt Nam'
        rate:
          type: array
          uniqueItems: false
          items: {}
        price:
          type: number
          example: 5euro
        opentime:
          type: string
          example: 9 AM
          format: date-time
        contact_number:
          type: string
          example: +84 384 568 983
        description:
          type: string
        discount_rate:
          type: number
        instagram:
          type: string
        created_at:
          type: string
          format: date-time
        update_at:
          type: string
          format: date-time
        location:
          type: object
          required:
            - longitude
            - latitude
          properties:
            longitude:
              type: number
              example: 105.83552
            latitude:
              type: number
              example: 21.01066
      required:
        - items_id
        - name
        - address
        - rate
        - price
        - opentime
        - contact_number
        - description
        - discount_rate
        - instagram
        - created_at
        - update_at
        - location
      examples:
        - items_id: '1'
          name: Bazooni Barbershop
          address: 'Asemakatu 15, 90100 Oulu'
          rate: []
          price: 5
          opentime: '10:00'
          contact_number: +84 384 568 983
          description: string
          discount_rate: 0
          instagram: string
          created_at: '2019-08-24T14:15:22Z'
          update_at: '2019-08-24T14:15:22Z'
          location:
            longitude: 105.83552
            latitude: 21.01066
...